Implements fixed joint.
This is realized through a S matrix of size (6,0).
Such size is not accepted at run-time but the related
code is not executed when the DOF of the joint is equal to zero.
The only modification at the algorithmic level is the
introduction of the condition and the multiplication of S by ddqi
not done at the matrix level.

In the current commit the references of simple_arm are not
updated correctly. But it includes a new fixed joint the Galilean frame.
The jacobian references files have to be checked.
No overhead are induced by this commit.